requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
pydantic==2.7.0
crewai==0.51.1
langchain==0.2.16
langchain-community==0.2.12
requests==2.32.3
python-dotenv==1.0.1
celery==5.3.6
redis==5.0.8
astral==2.2